CIterator assumes BoxStream as its iterator
